fix: cast any injury label to half

+++ rsna/model.py
@@ -97,7 +97,7 @@ class CombinedLoss(nn.Module):
         kidney, liver, spleen = F.softmax(kidney, dim=-1), F.softmax(liver, dim=-1), F.softmax(spleen, dim=-1)
         healthy = torch.cat([kidney[:, 0:1], liver[:, 0:1], spleen[:, 0:1]], dim=-1)
         any_injury, _ = torch.max(1 - healthy, keepdim=True, dim=-1)
-        any_loss = F.binary_cross_entropy(any_injury, labels[:, 11:12].float(), weight=labels[:, 11:12] * 6 + 1)
+        any_loss = F.binary_cross_entropy(any_injury, labels[:, 11:12].half(), weight=labels[:, 11:12] * 6 + 1)
         return ce_loss + any_loss
